Format code with rustfmt
This formats the codebase in preparation for a CI rustfmt check in #15. Nothing
else changes in this PR, it is only code formatting.

I was hoping to update `openapi.patch` to support the formatting, but the patch
was already pretty broken due to upstream schema and tool changes so that needs
to be updated manually regardless.

Connects #15

Merges #16

LGTM given by: @katiemckeon