Improve dev ergonomics
A handful of miscellaneous ergonomics improvements:

- Improve caching strategy for builds on CI. This brings the build time down
from 6 minutes and 46 seconds down to 1 minute and 50 seconds.
- Replaces the deprecated actions-rs actions which were causing CI warnings
with a modern alternative.
- Adds a `rustfmt.toml` config and checks formatting on CI.
- Adds a compose project for easier cross-platform development. This also pulls
in the non-rust dependencies needed for updating the openapi client, uses
caching volumes for fast builds, and sources a local `.bashrc` file for
customizations. I added a `.bashrc.example` file to persist the history in
the container and make it perform a little nicer.
- Locks the `@openapitools/openapi-generator-cli` version in the
`update-sdk.sh` script so it doesn't drift over time. I think part of the
reason the patch won't apply cleanly is because of changes in this tool
version. I also deleted another file that it's creating and allowed the
script to be run from anywhere.

Merges #15

LGTM given by: @katiemckeon